refactor(web): split settings-ui into one-component-per-file modules
Move SectionHeading, SettingRow, and SettingsSaveBar out of the shared
settings-ui.tsx into their own files and update the section imports.
