+++ apps/web/src/features/settings/components/section-heading.tsx
@@ -1,0 +1,14 @@
+export function SectionHeading({
+	title,
+	description,
+}: {
+	title: string;
+	description: string;
+}) {
+	return (
+		<div className="mb-6 space-y-1">
+			<h1 className="font-medium text-2xl tracking-tight">{title}</h1>
+			<p className="text-muted-foreground text-sm">{description}</p>
+		</div>
+	);
+}
